move canvas components into directory
