+++ src/components/Canvas/CanopySpaceView.tsx
