+++ src/components/Canvas/CanopyView.tsx
