+++ src/components/Canvas/PreviewCanvas.tsx
@@ -3,12 +3,12 @@ import { Canvas } from "@react-three/fiber";
 import { useStore } from "@/src/types/StoreContext";
 import { observer } from "mobx-react-lite";
 import { RenderingGate } from "@/src/components/RenderingGate";
-import { Canopy } from "@/src/components/Canopy";
+import { Canopy } from "@/src/components/Canvas/CanopyView";
 import { CameraControls } from "@/src/components/CameraControls";
 import { SingleBlockRenderPipeline } from "@/src/components/RenderPipeline/SingleBlockRenderPipeline";
 import { useState } from "react";
 import { WebGLRenderTarget } from "three";
-import { CanopySpaceView } from "@/src/components/CanopySpaceView";
+import { CanopySpaceView } from "@/src/components/Canvas/CanopySpaceView";
 
 type PreviewCanvasProps = {
   block: Block;
